Disable editing workunits from incidents.

+++ include/functions_form.php
@@ -864,7 +864,6 @@ echo "<form method='post' action='index.php?sec=projects&sec2=operation/projects
 	echo "</table>";
 	echo "</div>";
 
-
 	// Body
 	//echo "<div class='notebody'>";
 	$output = "<div class='notebody' id='wu_$id_workunit'>";
@@ -891,27 +890,33 @@ echo "<form method='post' action='index.php?sec=projects&sec2=operation/projects
 			$myurl = "index.php?sec=users&sec2=operation/users/user_workunit_report&id=$id_user";
 	}
 	
-	if ((project_manager_check($id_project) == 1) OR ($id_user == $config["id_user"]) OR  (give_acl($config["id_user"], 0, "TM")) ) {
+	$belong_to_ticket = get_db_value_sql("SELECT * FROM tworkunit_incident WHERE id_workunit = ".$id_workunit);
+	
+	//~ if ((project_manager_check($id_project) == 1) OR ($id_user == $config["id_user"]) OR  (give_acl($config["id_user"], 0, "TM")) ) {
+	if (((project_manager_check($id_project) == 1) OR ($id_user == $config["id_user"]) OR  (give_acl($config["id_user"], 0, "TM"))) && (!$belong_to_ticket)) {
 		$output .= "<tr><td align='right'>";
 		$output .= "<a class='delete-workunit' id='delete-$id_workunit' href='$myurl&id_workunit=$id_workunit&operation=delete' onclick='if (!confirm(\"".__('Are you sure?')."\")) return false;'><img src='images/cross.png'  title='".__('Delete workunit')."'/></a>";
 	}
 
 	// Edit workunit
-	if (((project_manager_check($id_project) == 1) OR (give_acl($config["id_user"], 0, "TM")) OR ($id_user == $config["id_user"])) AND (($locked == "") OR (give_acl($config["id_user"], 0, "UM")) )) {
+	//~ if (((project_manager_check($id_project) == 1) OR (give_acl($config["id_user"], 0, "TM")) OR ($id_user == $config["id_user"])) AND (($locked == "") OR (give_acl($config["id_user"], 0, "UM")) )) {
+	if ((((project_manager_check($id_project) == 1) OR (give_acl($config["id_user"], 0, "TM")) OR ($id_user == $config["id_user"])) AND (($locked == "") OR (give_acl($config["id_user"], 0, "UM")) )) && (!$belong_to_ticket)) {
 		$output .= "<tr><td align='right'>";
 		$output .= "<a class='edit-workunit' id='edit-$id_workunit' href='index.php?sec=projects&sec2=operation/users/user_spare_workunit&id_project=$id_project&id_task=$id_task&id_workunit=$id_workunit&id_profile=$id_profile'><img border=0 src='images/page_white_text.png' title='".__('Edit workunit')."'></a>";
 		$output .= "</td>";
 	}
 
 	// Lock workunit
-	if (((project_manager_check($id_project) == 1) OR (give_acl($config["id_user"], 0, "TM")) OR ($id_user == $config["id_user"])) AND (($locked == "")  )) {
-		$output .= "<tr><td align='right'>";
-		$output .= "<a class='lock_workunit' id='lock-$id_workunit' href='$myurl&id_workunit=$id_workunit&operation=lock'><img src='images/lock.png' title='".__('Lock workunit')."'></a>";
-		$output .= "</td>";
-	} else {
-		$output .= "<tr><td align='right'>";
-		$output .= "<img src='images/rosette.png' title='".__('Locked by')." $locked'";
-		$output .= "</td>";
+	if (!$belong_to_ticket) {
+		if (((project_manager_check($id_project) == 1) OR (give_acl($config["id_user"], 0, "TM")) OR ($id_user == $config["id_user"])) AND (($locked == ""))) {
+			$output .= "<tr><td align='right'>";
+			$output .= "<a class='lock_workunit' id='lock-$id_workunit' href='$myurl&id_workunit=$id_workunit&operation=lock'><img src='images/lock.png' title='".__('Lock workunit')."'></a>";
+			$output .= "</td>";
+		} else {
+			$output .= "<tr><td align='right'>";
+			$output .= "<img src='images/rosette.png' title='".__('Locked by')." $locked'";
+			$output .= "</td>";
+		}
 	}
 
   	$output .= "</tr></table>";
